+++ packages/template-retail-react-app/translations/en-GB.json
@@ -122,9 +122,6 @@
   "account_order_detail.alert.return_success_title": {
     "defaultMessage": "Return submitted"
   },
-  "account_order_detail.badge.cancelled": {
-    "defaultMessage": "Cancelled"
-  },
   "account_order_detail.button.cancel_order": {
     "defaultMessage": "Cancel order"
   },
@@ -214,9 +211,6 @@
   },
   "account_order_detail.title.order_details": {
     "defaultMessage": "Order Details"
-  },
-  "account_order_history.badge.cancelled": {
-    "defaultMessage": "Cancelled"
   },
   "account_order_history.button.continue_shopping": {
     "defaultMessage": "Continue Shopping"
@@ -1388,6 +1382,9 @@
   },
   "offline_banner.description.browsing_offline_mode": {
     "defaultMessage": "You're currently browsing in offline mode"
+  },
+  "order_status_badge.label.cancelled": {
+    "defaultMessage": "Cancelled"
   },
   "order_summary.action.remove_promo": {
     "defaultMessage": "Remove"
